requirements.txt
accelerate==1.2.0
pillow==10.2.0
numpy==1.26.3
pandas==2.2.3
wandb==0.19.0
tqdm==4.67.1
scikit-learn==1.5.2
vit-pytorch==1.8.9

# instalar com cuda
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1